Moved many layer queries into storage traits.
The reason for doing this is to ensure that we can accurately query
layer information, even when the layer is not loaded. This is the case
when we're dealing with a delta rollup layer. The rollup layer itself
will not be able to answer questions like 'how many triples were added
in this layer' or 'what are all the additions for this particular
object', since it is an aggregate of many layers. For these queries,
we need to look at the actual layer on disk. This refactor does so,
but hides that this is done in the case of the StoreLayer, which
carries with it a reference to the layerstore it came out of. For
the CachedLayerStore, we check if the layer is in memory before going
to disk, allowing fast layer querying still for the non-rollup case.