Speed improvement in LogArray::entry for non-boundary cases (#95)
In a logarray, an entry may be entirely within one aligned u64, or it may span
two. In the existing code, we always retrieve two u64's, but will only use one
if the entry doesn't span both. This change will only retrieve that second u64
when it is actually needed.

It's a small change, and the speed improvement isn't huge, but LogArray::entry
appears to be the most called function in TerminusDB by a large margin, so even
small improvements matter a little.