Pre-allocate vectors instead of growing them dynamically (#97)
Various places in the code were allocating vectors that'd have to be grown at a
later point, which could incur a full move of the existing array data to a new
location in memory, incurring quite some overhead. This is now avoided by
ensuring vectors are pre-allocated with the right size.