Introduce method for quickly looking up a single triple when subject and predicate are known (#98)
There are some cases where we very quickly want to look up a single triple based
on a subject and a predicate. A major place where this comes up is when type
checking, as we always expect just a single triple to match `t(id, rdf:type,
Type)`. Other examples are list cells where we always expect a single
`rdf:first` and `rdf:rest`, or arrays where we always expect a single
`sys:index` and `sys:value`.

Avoiding setting up a full iterator in those cases makes code go faster. So even
though it's a bit of a special case, having a function just for this makes
sense.